Update Svg component
