+++ src/ui/Svg/index.tsx
@@ -1,0 +1,1 @@
+export { Svg } from './Svg'
