--- src/ui/svg/HeaderSvg/HeaderSvg.tsx
@@ -1,57 +1,0 @@
-import { FC } from 'react'
-
-import { INameIcons } from '@/types/Icons'
-
-interface HeaderSvgProps {
-  size?: 'small' | 'middle' | 'big'
-  icon: INameIcons
-}
-
-export const HeaderSvg: FC<HeaderSvgProps> = ({ size = 'middle', icon }) => {
-  const sz = {
-    small: 16,
-    middle: 20,
-    big: 32,
-  }
-
-  return (
-    <svg
-      version='1.1'
-      xmlns='http://www.w3.org/2000/svg'
-      width={sz[size]}
-      height={sz[size]}
-      viewBox='0 0 32 32'
-    >
-      {icon === 'call' && (
-        <path d='M7.111 4q-0.889 0-1.778 0.889v0q-0.444 0.444-0.889 1.333v0q-0.444 0.889-0.444 2.444t0.889 3.556 2.444 4.222 3.778 4.444 4.444 3.778 4.222 2.444 3.556 0.889 2.444-0.444v0q0.889-0.444 1.333-0.889l0.889-1.333v-0.889l-5.333-3.556-1.778 1.333q-0.889 0.889-2.222 0.889l-1.333-0.444q-0.889-0.444-2.222-1.333l-4.444-4.444-1.333-2.222-0.444-1.778q0-0.889 0.444-1.778l1.778-1.778-3.556-5.333h-0.444zM2.667 1.778q1.778-1.778 4.444-1.778 0.889 0 2 0.444t1.556 1.333l4 6.222q0.444 0.444 0.444 1.556t-0.444 2l-1.778 1.778 2.667 3.111 3.111 2.667 1.778-1.778q0.889-0.444 1.778-0.444t1.778 0.444l1.333 0.444 4.889 3.556q0.889 0.444 1.333 1.778v0q0.444 0.889 0.444 2t-0.444 2.222-1.556 2.222-2.889 1.778-4 0.667-4.889-1.111-5.333-2.889v0l-4.444-4.444q-2.667-2.222-4.444-4.889v0q-1.778-2.222-2.667-4.889v0q-1.333-2.667-1.333-4.889t0.667-4 2-3.111z'></path>
-      )}
-      {icon === 'copy' && (
-        <path d='M22.222 0q1.778 0 3.778 1.111t2.889 2.889 0.889 4v12q0.444 0.889-0.222 1.333t-1.556 0.444-1.333-0.444-0.444-1.333v-12q-0.444-1.778-1.556-2.889t-2.444-1.111h-8q-0.889 0-1.556-0.667t-0.667-1.333 0.667-1.333 1.111-0.667h8.444zM2.222 13.778q0-2.222 1.556-4t4.222-1.778h8q2.667 0 4.222 1.778t2 4v12q-0.444 2.667-2 4.444t-4.222 1.778h-8q-2.667 0-4.222-1.778t-1.556-4.444v-12zM8 12q-0.889 0-1.333 0.667t-0.444 1.111v12q0 0.889 0.444 1.556t1.333 0.667h8q0.889 0 1.556-0.667t0.667-1.111v-12.444q0-0.444-0.667-1.111t-1.556-0.667h-8z'></path>
-      )}
-      {icon === 'favoriteAdd' && (
-        <path d='M12.819 24.679l0.224-0.128q1.375-0.767 2.957-0.767t2.957 0.767l0.224 0.128 6.426 4.092v-25.542q-0.032-0.032 0-0.032h-19.213q0.032 0 0 0.032v25.479zM6.394 0h19.213q1.311 0 2.254 0.975t0.943 2.318v25.415q0 1.343-0.943 2.318t-2.254 0.975q-0.799 0-1.471-0.384l-6.681-4.252q-0.671-0.384-1.455-0.384t-1.455 0.384l-6.681 4.252q-1.183 0.639-2.446 0.224t-1.87-1.63q-0.352-0.703-0.352-1.503v-25.415q0-1.343 0.943-2.318t2.254-0.975z'></path>
-      )}
-      {icon === 'favoriteRemove' && (
-        <path d='M6.394 0h19.213q1.311 0 2.254 0.975t0.943 2.318v25.415q0 1.343-0.943 2.318t-2.254 0.975q-0.799 0-1.471-0.384l-6.681-4.252q-0.671-0.384-1.455-0.384t-1.455 0.384l-6.681 4.252q-1.183 0.639-2.446 0.224t-1.87-1.63q-0.352-0.703-0.352-1.503v-25.415q0-1.343 0.943-2.318t2.254-0.975z'></path>
-      )}
-      {icon === 'language' && (
-        <path d='M16 32q-4.444 0-8-2.222t-5.778-5.778-2.222-8 2.222-8 5.778-5.778 8-2.222 8 2.222 5.778 5.778 2.222 8-2.222 8-5.778 5.778-8 2.222zM14.667 21.333v-4h-5.333q0 2.222 0.889 4.444 2.222-0.444 4.444-0.444zM14.667 24q-1.778 0-3.556 0.444l0.444 1.333q1.333 2.667 3.111 3.111v-4.889zM7.556 22.667q-0.889-2.667-0.889-5.333h-4q0.444 3.556 2.667 6.222zM7.111 25.778q0.889 1.333 2.667 1.778l-1.333-2.222zM17.333 24v4.889q1.778-0.444 3.111-3.111l0.444-1.333q-1.778-0.444-3.556-0.444zM21.778 21.778q0.889-2.222 0.889-4.444h-5.333v4q2.222 0 4.444 0.444zM23.556 25.333q-0.444 1.333-1.333 2.222 1.333-0.444 2.667-1.778zM26.667 24q2.222-3.111 2.667-6.667h-4q0 2.667-0.889 5.333zM9.333 14.667h5.333v-4q-2.222 0-4.444-0.444-0.889 2.222-0.889 4.444zM11.111 7.556q1.778 0.444 3.556 0.444v-4.889q-1.778 0.444-3.111 3.111zM17.333 14.667h5.333q0-2.222-0.889-4.444-2.222 0.444-4.444 0.444v4zM17.333 8q1.778 0 3.556-0.444l-0.444-1.333q-1.333-2.667-3.111-3.111v4.889zM25.333 14.667h4q-0.444-3.556-2.667-6.667l-2.222 1.333q0.889 2.667 0.889 5.333zM23.556 6.667q0.889 0 1.333-0.444-1.333-1.333-2.667-1.778zM9.778 4.444q-1.778 0.444-2.667 1.778l1.333 0.444zM2.667 14.667h4q0-2.667 0.889-5.333l-2.222-0.889q-2.222 2.667-2.667 6.222z'></path>
-      )}
-      {icon === 'notify' && (
-        <path d='M4.889 11.111q0-3.111 1.556-5.556t4-4 5.556-1.556 5.556 1.556 4 4 1.556 5.556v10.667l0.444 0.889 2.222 2.222q0.444 0.889 0.444 1.778t-0.667 1.556-1.556 0.667h-7.556q-0.444 1.333-1.778 2.222t-2.667 0.889-2.667-0.889-1.778-2.222h-7.556q-0.889 0-1.556-0.667t-0.667-1.556 0.444-1.778l2.222-2.222 0.444-0.889v-10.667zM23.556 22.667q-0.444-0.889-0.444-2.222v-9.333q0-1.778-0.889-3.556t-2.667-2.667-3.556-0.889-3.556 0.889-2.667 2.667-0.889 3.556v9.333l-0.444 2.222q0 1.333-0.889 2.222h16.889q-0.889-0.889-0.889-2.222z'></path>
-      )}
-      {icon === 'playOutline' && (
-        <path d='M6.222 4v24l18.667-12zM25.778 15.556v0 0zM3.111 2.667q0-0.889 0.444-1.556t1.556-0.889 2 0.222l20.444 13.333q0.444 0.444 0.889 1.111t0.222 1.778-1.111 1.556l-20.444 13.333q-0.889 0.444-1.778 0.444t-1.556-0.667-0.667-2v-26.667z'></path>
-      )}
-      {icon === 'profile' && (
-        <path d='M11.111 8q0-1.778 1.556-3.333t3.333-1.556 3.333 1.556 1.556 3.333-1.556 3.333-3.333 1.556-3.333-1.556-1.556-3.333zM16 0q-2.222 0-4 1.111t-2.889 2.889-1.111 4 1.111 4 2.889 2.889 4 1.111 4-1.111 2.889-2.889 1.111-4-1.111-4-2.889-2.889-4-1.111zM7.556 24q3.556-1.778 8.444-1.778t8.444 1.778q1.333 1.333 1.778 2.667l0.889 2.667q0.444 0.444 1.111 0.889t1.111 0.222 0.889-0.889 0-1.111l-0.889-2.667q-0.444-1.333-1.333-2.667t-2.222-1.778q-4-2.222-9.778-2.222t-9.778 2.222q-1.333 0.444-2.222 1.778t-1.333 2.667l-0.889 2.667q-0.444 0.444 0 1.111t0.889 0.889 1.111-0.222 1.111-0.889l0.889-2.667q0.444-1.333 1.778-2.667z'></path>
-      )}
-      {icon === 'search' && (
-        <path d='M25.333 22.222q0.444 0 0.444 0.444l5.333 5.333q0.889 0.889 0.889 1.778t-0.667 1.556-1.556 0.667-1.778-0.444l-5.778-6.222q-3.556 2.667-7.778 2.667t-7.556-1.778-5.111-5.111-1.778-7.111 1.778-7.111 5.111-5.111 7.111-1.778 7.111 1.778 5.111 5.111 1.778 7.556-2.667 7.778zM14.222 4q-3.111 0-5.333 1.333t-3.556 3.556-1.333 5.111 1.333 5.111 3.556 3.556 5.111 1.333 5.111-1.333 3.556-3.556 1.333-5.111-1.333-5.111-3.556-3.556-4.889-1.333z'></path>
-      )}
-      {icon === 'share' && (
-        <path d='M17.778 5.333v15.556q0 0.444-0.667 0.889t-1.111 0.444-1.111-0.444-0.667-0.889v-15.556l-1.778 2.222q-0.444 0.444-1.111 0.444t-1.111-0.444-0.444-1.111 0.444-1.111l4.444-4.889q0.889-0.444 1.333-0.444t1.333 0.444l4.444 4.889q0.444 0.444 0.444 1.111t-0.444 1.111-1.111 0.444-1.111-0.444zM3.111 16q0.889 0 1.333 0.444t0.444 1.333v8l0.444 1.333q0.444 0.889 1.333 1.333l1.333 0.444h16l1.333-0.444q0.889-0.444 1.333-1.333l0.444-1.333v-8q0-0.889 0.444-1.333t1.111-0.444 1.111 0.444 0.444 1.333v8.444q0 1.333-0.667 2.667t-2.222 2.222-2.889 0.889h-16.889q-1.333 0-2.889-0.889t-2.222-2.222-0.667-2.667v-8.444q0-0.889 0.444-1.333t0.889-0.444z'></path>
-      )}
-    </svg>
-  )
-}
