--- src/ui/svg/HeaderSvg/index.tsx
@@ -1,1 +1,0 @@
-export {HeaderSvg} from './HeaderSvg'
